+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+##0.0.54
+
+* Set taken picture size equal to preview size in CameraAPI 1
+
+##0.0.53
+
+* Avoid take picture before starting preview session completely
+##0.0.52
+
+* Orientation in CameraAPI 1 fixed
+
+##0.0.50
+
+* Support from CameraAPI 1 now, for taking picture with flash in old phones like samsung S6
 
 ##0.0.49